Add ability to use where/bind clauses in LogBrowser
Loading in all logs associated with a given task in a given collection
can take a very long time (i.e. depending on how many dataRefs exist
for that task.)  This adds the ability to add a "where" clause to the
dataId search such that the log loading time can be greatly reduced.
A "bind" mapping can (optionally) be used with the where clause.  If
a given key provided in the bind mapping does not appear in the where
clause string, a warning will be logged (stating that no binding will
take effect on that key).